turtle art using math
i made an equation that generates massive numbers from small numbers, and uses the answer to the equation to draw shapes. each line's length and direction is based on each digit of the answer. there are 2 types of functions that draw. one is more scribbly as the directional change's angle is determined by the digit, whereas the other simply sticks to 90 degree angle changes, the direction of which is determined by if the digit is even or odd.

not sure what this is good for but it was fun and it creates some cool lookin visuals