turtle collatz visual display
a visual display of the collatz conjecture, showing the length of steps it takes for a given value to get to 1 using the conjecture. it takes a value, passes it through the conjecture while keeping track of how many times it passes through the collatz() func, and once it reaches 1 it takes the value of steps and uses that to draw a kind of line graph showing how many steps it took in pixels for each value. this one is by far my favorite turtle project so far, i love playing around with the collatz conjecture